requirements.txt
fastapi[standard]
beanie
motor
pydantic-settings
